# docs/swagger.yaml
basePath: /api
definitions:
  controllers.ParameterInfo:
    properties:
      symbols:
        items:
          $ref: '#/definitions/controllers.ParameterProperty'
        type: array
    type: object
  controllers.ParameterProperty:
    properties:
      parameter:
        type: string
      property:
        type: string
    type: object
  dynamics.IterationResult:
    properties:
      impacts:
        items:
          $ref: '#/definitions/impact.Impact'
        type: array
    type: object
  impact.Impact:
    properties:
      phase:
        type: number
      time:
        type: number
      velocity:
        type: number
    type: object
info:
  contact: {}
  description: Analysis and simulation of a simple vibro-impact model developed in
    Go - principally as a learning exercise
  title: Impact Oscillator
  version: "1.0"
paths:
  /iteration/data/:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Return data from iterating the impact map for a specified set of
        parameters
      operationId: post-iteration-data
      parameters:
      - description: Forcing frequency
        in: formData
        minimum: 0
        name: frequency
        required: true
        type: number
      - description: Obstacle offset from origin
        in: formData
        name: offset
        required: true
        type: number
      - description: Coefficient of restitution
        in: formData
        maximum: 1
        minimum: 0
        name: r
        required: true
        type: number
      - default: 100
        description: Number of periods without an impact after which the algorithm
          will report 'long excursions'
        in: formData
        name: maxPeriods
        type: integer
      - default: 0
        description: Phase at initial impact
        in: formData
        name: phi
        required: true
        type: number
      - default: 0
        description: Velocity at initial impact
        in: formData
        name: v
        required: true
        type: number
      - default: 5000
        description: Number of iterations of impact map
        in: formData
        name: numIterations
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dynamics.IterationResult'
        "400":
          description: Invalid parameters
          schema:
            type: string
      summary: Impact data
  /iteration/image/:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Return scatter plot from iterating the impact map for a specified
        set of parameters
      operationId: post-iteration-image
      parameters:
      - description: Forcing frequency
        in: formData
        minimum: 0
        name: frequency
        required: true
        type: number
      - description: Obstacle offset from origin
        in: formData
        name: offset
        required: true
        type: number
      - description: Coefficient of restitution
        in: formData
        maximum: 1
        minimum: 0
        name: r
        required: true
        type: number
      - default: 100
        description: Number of periods without an impact after which the algorithm
          will report 'long excursions'
        in: formData
        name: maxPeriods
        type: integer
      - default: 0
        description: Phase at initial impact
        in: formData
        name: phi
        required: true
        type: number
      - default: 0
        description: Velocity at initial impact
        in: formData
        name: v
        required: true
        type: number
      - default: 5000
        description: Number of iterations of impact map
        in: formData
        name: numIterations
        type: integer
      produces:
      - image/png
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dynamics.IterationResult'
        "400":
          description: Invalid parameters
          schema:
            type: string
      summary: Impact map
  /parameter-info/groups/:
    get:
      description: Groups for displaying related parameters
      operationId: get-parameter-groups
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.ParameterInfo'
      summary: Parameter groups
  /parameter-info/symbols/:
    get:
      description: Greek symbols to be used for rendering specified parameters
      operationId: get-parameter-symbols
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.ParameterInfo'
      summary: Parameter symbols
  /singularity-set/data/:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Return impacts which map to and from zero velocity impacts for
        a specified set of parameters
      operationId: post-singularity-set-data
      parameters:
      - description: Forcing frequency
        in: formData
        minimum: 0
        name: frequency
        required: true
        type: number
      - description: Obstacle offset from origin
        in: formData
        name: offset
        required: true
        type: number
      - description: Coefficient of restitution
        in: formData
        maximum: 1
        minimum: 0
        name: r
        required: true
        type: number
      - default: 100
        description: Number of periods without an impact after which the algorithm
          will report 'long excursions'
        in: formData
        name: maxPeriods
        type: integer
      - default: 5000
        description: Number of impacts to map
        in: formData
        name: numPoints
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dynamics.IterationResult'
        "400":
          description: Invalid parameters
          schema:
            type: string
      summary: Singularity set data
  /singularity-set/image/:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Return scatter plot of impacts which map to and from zero velocity
        impacts for a specified set of parameters
      operationId: post-singularity-set-image
      parameters:
      - description: Forcing frequency
        in: formData
        minimum: 0
        name: frequency
        required: true
        type: number
      - description: Obstacle offset from origin
        in: formData
        name: offset
        required: true
        type: number
      - description: Coefficient of restitution
        in: formData
        maximum: 1
        minimum: 0
        name: r
        required: true
        type: number
      - default: 100
        description: Number of periods without an impact after which the algorithm
          will report 'long excursions'
        in: formData
        name: maxPeriods
        type: integer
      - default: 5000
        description: Number of impacts to map
        in: formData
        name: numPoints
        type: integer
      produces:
      - image/png
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dynamics.IterationResult'
        "400":
          description: Invalid parameters
          schema:
            type: string
      summary: Singularity set
swagger: "2.0"